app/requirements.txt
fastapi
uvicorn 
python-dotenv
alembic
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic[email] 
python-multipart
pandas
python-jose[cryptography]
passlib[bcrypt]
openpyxl
pymongo